+++ src/Data/Serialize/Describe/Combinators/Byte.hs
@@ -4,9 +4,10 @@ import Data.Word
 import Data.Int
 import Data.Serialize.Describe.Descriptor
 import Data.Serialize.Describe.Class
+import Data.Serialize.Describe.Isomorphisms
 
 w8 :: Integral i => (s -> i) -> Descriptor s i
-w8 f = fromIntegral <$> describe @Word8 (fromIntegral . f)
+w8 = isoField @Word8 fi
 
 i8 :: Integral i => (s -> i) -> Descriptor s i
-i8 f = fromIntegral <$> describe @Int8 (fromIntegral . f)
+i8 = isoField @Int8 fi